chore: rename infection-and-mortality-rates to infection-and-mortality-data
