+++ src/components/pages/race/infection-and-mortality-data/share.module.scss
