+++ src/templates/race/infection-and-mortality-rates/landing.js
@@ -35,7 +35,7 @@ export default ({ pageContext, path, data }) => {
       returnLinks={[
         { link: '/race' },
         {
-          link: '/race/infection-and-mortality-rates',
+          link: '/race/infection-and-mortality-data',
           title: 'Infection and Mortality Rates',
         },
       ]}
@@ -75,7 +75,7 @@ export default ({ pageContext, path, data }) => {
           </Link>
         </li>
         <li>
-          <Link to="/race/infection-and-mortality-rates">
+          <Link to="/race/infection-and-mortality-data">
             Infection and Mortality Rates by Race and Ethnicity
           </Link>
         </li>
